Split textbox mode memory tests
